requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
python-whois==0.8.0
dnspython==2.4.2
geoip2==4.7.0
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
aioredis==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
aioresponses==0.7.6
